refactor: Update context imports and streamline user access hooks
- Refactored multiple context files to utilize a centralized hooks structure, improving organization and maintainability.
- Updated imports across various components to reflect the new hooks directory, enhancing clarity and reducing redundancy.
- Removed obsolete context definitions and custom hooks, streamlining the codebase for better performance and readability.
