+++ src/contexts/ServiceContext.tsx
@@ -1,31 +1,12 @@
 "use client";
 
-import { ApiService } from "@/domain/interfaces/api/ApiService";
-import { MemberService } from "@/domain/interfaces/services/MemberService";
-import { PusherService } from "@/domain/interfaces/services/PusherService";
 import { ApiServiceImpl } from "@/services/api/ApiServiceImpl";
 import { MockApiService } from "@/services/api/MockApiService";
 import { MemberServiceImpl } from "@/services/member/MemberServiceImpl";
 import { MockMemberService } from "@/services/member/MockMemberService";
 import { PusherServiceImpl } from "@/services/pusher/PusherServiceImpl";
-import { createContext, ReactNode, useContext } from "react";
-
-/**
- * ServiceContainer interface defines the shape of our service container.
- * All services should be added to this interface.
- */
-export interface ServiceContainer {
-  apiService: ApiService;
-  memberService: MemberService;
-  pusherService: PusherService;
-  // Add more services here as they're created
-}
-
-/**
- * Create the context with a default undefined value.
- * This forces consumers to use the useServices hook which performs a null check.
- */
-const ServiceContext = createContext<ServiceContainer | undefined>(undefined);
+import { ReactNode } from "react";
+import { ServiceContext, ServiceContainer } from "./service.types";
 
 /**
  * Determine if we should use mocks based on environment or query params
@@ -97,43 +78,4 @@ export function ServiceProvider({
   return (
     <ServiceContext.Provider value={value}>{children}</ServiceContext.Provider>
   );
-}
-
-/**
- * Custom hook to access the services from the ServiceContext.
- * Throws an error if used outside of a ServiceProvider.
- */
-export function useServices(): ServiceContainer {
-  const context = useContext(ServiceContext);
-
-  if (context === undefined) {
-    throw new Error("useServices must be used within a ServiceProvider");
-  }
-
-  return context;
-}
-
-/**
- * Convenience hook to access just the ApiService.
- * This pattern can be repeated for other services to avoid destructuring.
- */
-export function useApiService(): ApiService {
-  const { apiService } = useServices();
-  return apiService;
-}
-
-/**
- * Convenience hook to access just the MemberService.
- */
-export function useMemberService(): MemberService {
-  const { memberService } = useServices();
-  return memberService;
-}
-
-/**
- * Convenience hook to access just the PusherService.
- */
-export function usePusherService(): PusherService {
-  const { pusherService } = useServices();
-  return pusherService;
 }
