+++ src/contexts/UserAccessContext.tsx
@@ -6,47 +6,8 @@ import { MeteredUsage } from "@/domain/entities/MeteredUsage";
 import { useAuth } from "@/hooks/auth";
 import { useLicenseService } from "@/hooks/useLicenseService";
 import { useMeteredUsageService } from "@/hooks/useMeteredUsageService";
-import React, {
-  createContext,
-  useContext,
-  useState,
-  useEffect,
-  ReactNode,
-  useCallback,
-} from "react";
-
-interface UserAccessContextType {
-  activeLicenses: License[];
-  licensePolicies: LicensePolicy[];
-  meteredUsage: MeteredUsage[];
-  isLoading: boolean; // Combined loading state
-  canAccessFeature: (feature: string) => boolean;
-  isMeterLimitReached: (meterId: MeteredFeature) => boolean;
-  getTotalAggregatedValueForMeter: (meterId: MeteredFeature) => number;
-  getUsageForMeter: (meterId: MeteredFeature) => MeteredUsage | undefined;
-  getHighestLimitForMeter: (meterId: MeteredFeature) => number;
-  refreshAccessData: () => Promise<void>; // Combined refresh
-  isLoaded: boolean;
-}
-
-const defaultContextValue: UserAccessContextType = {
-  activeLicenses: [],
-  licensePolicies: [],
-  meteredUsage: [],
-  isLoading: true,
-  canAccessFeature: () => false,
-  isMeterLimitReached: () => true, // Default to true (restricted) while loading or if no data
-  getTotalAggregatedValueForMeter: () => 0,
-  getUsageForMeter: () => undefined,
-  getHighestLimitForMeter: () => 0,
-  refreshAccessData: async () => {},
-  isLoaded: false,
-};
-
-const UserAccessContext =
-  createContext<UserAccessContextType>(defaultContextValue);
-
-export const useUserAccessContext = () => useContext(UserAccessContext);
+import React, { useState, useEffect, ReactNode, useCallback } from "react";
+import { UserAccessContext, UserAccessContextType } from "./user-access.types";
 
 interface UserAccessProviderProps {
   children: ReactNode;
