+++ src/contexts/VerificationContext.tsx
@@ -1,23 +1,8 @@
 "use client";
 
-import { createContext, useContext, ReactNode, useMemo } from "react";
+import { ReactNode, useMemo } from "react";
 import { useVerificationService } from "@/hooks/useVerificationService";
-
-/**
- * VerificationState interface defines the shape of our verification context value.
- */
-export interface VerificationState {
-  isEmailVerified: boolean;
-  isUserVerified: boolean;
-}
-
-/**
- * Create the context with a default undefined value.
- * This forces consumers to use the useVerification hook which performs a null check.
- */
-const VerificationContext = createContext<VerificationState | undefined>(
-  undefined
-);
+import { VerificationContext, VerificationState } from "./verification.types";
 
 interface VerificationProviderProps {
   children: ReactNode;
@@ -43,36 +28,4 @@ export function VerificationProvider({ children }: VerificationProviderProps) {
       {children}
     </VerificationContext.Provider>
   );
-}
-
-/**
- * Custom hook to access the verification state from the VerificationContext.
- * Throws an error if used outside of a VerificationProvider.
- */
-export function useVerification(): VerificationState {
-  const context = useContext(VerificationContext);
-
-  if (context === undefined) {
-    throw new Error(
-      "useVerification must be used within a VerificationProvider"
-    );
-  }
-
-  return context;
-}
-
-/**
- * Convenience hook to check if the user's email is verified
- */
-export function useIsEmailVerified(): boolean {
-  const { isEmailVerified } = useVerification();
-  return isEmailVerified;
-}
-
-/**
- * Convenience hook to check if the user is verified through any means
- */
-export function useIsUserVerified(): boolean {
-  const { isUserVerified } = useVerification();
-  return isUserVerified;
 }
