+++ src/contexts/stripe.types.ts
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+import { Stripe } from "@stripe/stripe-js";
+
+export interface StripeContextType {
+  stripe: Stripe | null;
+  isLoading: boolean;
+  error: Error | null;
+}
+
+export const StripeContext = createContext<StripeContextType | undefined>(
+  undefined
+);
